# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/v1/pod/{name}/logs/ws:
    get:
      consumes:
      - application/json
      description: Stream real-time pod logs via WebSocket connection with support
        for multiple containers, previous logs, and filtering
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      - description: Pod name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name
        in: query
        name: cluster
        type: string
      - description: Container name (defaults to first container)
        in: query
        name: container
        type: string
      - description: Stream logs from all containers
        in: query
        name: all-containers
        type: boolean
      - description: Include logs from previous pod instance
        in: query
        name: previous
        type: boolean
      - description: Get all logs (ignores tail-lines)
        in: query
        name: all-logs
        type: boolean
      - description: 'Number of lines to tail (default: 100)'
        in: query
        name: tail-lines
        type: integer
      - description: Start time for logs (RFC3339 format)
        in: query
        name: since-time
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: WebSocket connection established
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Pod not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Stream Pod Logs via WebSocket
      tags:
      - WebSocket
...
swagger: "2.0"